// packages/builder/src/core/packageJson.d.ts
type JsonObject = { [Key in string]: JsonValue } & {
  [Key in string]?: JsonValue | undefined;
};

/**
Matches a JSON array.
*/
type JsonArray = JsonValue[] | readonly JsonValue[];

/**
Matches any valid JSON primitive value.
*/
type JsonPrimitive = string | number | boolean | null;

/**
Matches any valid JSON value.

@see `Jsonify` if you need to transform a type to one that is assignable to `JsonValue`.
*/
type JsonValue = JsonPrimitive | JsonObject | JsonArray;
// ...
declare namespace PackageJson {
// ...
  export type DirectoryLocations = {
    [directoryType: string]: JsonValue | undefined;

    /**
		Location for executable scripts. Sugar to generate entries in the `bin` property by walking the folder.
		*/
    bin?: string;

    /**
		Location for Markdown files.
		*/
    doc?: string;

    /**
		Location for example scripts.
		*/
    example?: string;

    /**
		Location for the bulk of the library.
		*/
    lib?: string;

    /**
		Location for man pages. Sugar to generate a `man` array by walking the folder.
		*/
    man?: string;

    /**
		Location for test files.
		*/
    test?: string;
  };
// ...
  export type Dependency = Partial<Record<string, string>>;

  /**
	A mapping of conditions and the paths to which they resolve.
	*/
  type ExportConditions = {
    [condition: string]: Exports;
  };

  /**
	Entry points of a module, optionally with conditions and subpath exports.
	*/
  export type Exports =
    | null
    | string
    | Array<string | ExportConditions>
    | ExportConditions;

  /**
	Import map entries of a module, optionally with conditions and subpath imports.
	*/
  export type Imports = {
    [key: `#${string}`]: Exports;
  };

  export interface NonStandardEntryPoints {
    /**
		An ECMAScript module ID that is the primary entry point to the program.
		*/
    module?: string;

    /**
		A module ID with untranspiled code that is the primary entry point to the program.
		*/
    esnext?:
      | string
      | {
          [moduleName: string]: string | undefined;
          main?: string;
          browser?: string;
        };

    /**
		A hint to JavaScript bundlers or component tools when packaging modules for client side use.
		*/
    browser?: string | Partial<Record<string, string | false>>;

    /**
		Denote which files in your project are "pure" and therefore safe for Webpack to prune if unused.

		[Read more.](https://webpack.js.org/guides/tree-shaking/)
		*/
    sideEffects?: boolean | string[];
  }

  export type TypeScriptConfiguration = {
    /**
		Location of the bundled TypeScript declaration file.
		*/
    types?: string;

    /**
		Version selection map of TypeScript.
		*/
    typesVersions?: Partial<Record<string, Partial<Record<string, string[]>>>>;

    /**
		Location of the bundled TypeScript declaration file. Alias of `types`.
		*/
    typings?: string;
  };
// ...
  export type YarnConfiguration = {
    /**
		If your package only allows one version of a given dependency, and you’d like to enforce the same behavior as `yarn install --flat` on the command-line, set this to `true`.

		Note that if your `package.json` contains `"flat": true` and other packages depend on yours (e.g. you are building a library rather than an app), those other packages will also need `"flat": true` in their `package.json` or be installed with `yarn install --flat` on the command-line.
		*/
    flat?: boolean;

    /**
		Selective version resolutions. Allows the definition of custom package versions inside dependencies without manual edits in the `yarn.lock` file.
		*/
    resolutions?: Dependency;
  };

  export type JSPMConfiguration = {
    /**
		JSPM configuration.
		*/
    jspm?: PackageJson;
  };

  /**
	Type for [npm's `package.json` file](https://docs.npmjs.com/creating-a-package-json-file). Containing standard npm properties.
	*/

  export interface PackageJsonStandard {
// ...
    exports?: Exports;

    /**
		Subpath imports to define internal package import maps that only apply to import specifiers from within the package itself.

		[Read more.](https://nodejs.org/api/packages.html#subpath-imports)
		*/
    imports?: Imports;

    /**
		The executable files that should be installed into the `PATH`.
		*/
    bin?: string | Partial<Record<string, string>>;
// ...
  }

  /**
	Type for [`package.json` file used by the Node.js runtime](https://nodejs.org/api/packages.html#nodejs-packagejson-field-definitions).
	*/
  export type NodeJsStandard = {
    /**
		Defines which package manager is expected to be used when working on the current project. It can set to any of the [supported package managers](https://nodejs.org/api/corepack.html#supported-package-managers), and will ensure that your teams use the exact same package manager versions without having to install anything else than Node.js.

		__This field is currently experimental and needs to be opted-in; check the [Corepack](https://nodejs.org/api/corepack.html) page for details about the procedure.__

		@example
		```json
		{
			"packageManager": "<package manager name>@<version>"
		}
		```
		*/
    packageManager?: string;
  };
// ...
}

/**
Type for [npm's `package.json` file](https://docs.npmjs.com/creating-a-package-json-file). Also includes types for fields used by other popular projects, like TypeScript and Yarn.
*/
export type PackageJson = JsonObject &
  PackageJson.NodeJsStandard &
  PackageJson.PackageJsonStandard &
  PackageJson.NonStandardEntryPoints &
  PackageJson.TypeScriptConfiguration &
  PackageJson.YarnConfiguration &
  PackageJson.JSPMConfiguration;